--- a/benchmark/benchmark.exs
+++ b/benchmark/benchmark.exs
@@ -1,6 +1,8 @@
 defmodule SWBenchmark do
   require Logger
 
+  @methods ~w(lists op_lists)a
+
   def get_texts(file_name) do
     lines = File.read!("benchmark/#{file_name}.txt")
       |> String.split("\n")
@@ -15,12 +17,12 @@ defmodule SWBenchmark do
     lines
   end
 
-  def compare_all_lines(lines) do
+  def compare_all_lines(lines, method) do
     lines
     |> Enum.each(fn l1 ->
       lines
       |> Enum.each(fn l2 ->
-        SmithWaterman.score_ratio(l1, l2)
+        SmithWaterman.score_ratio(l1, l2, 0.7, method)
       end)
     end)
   end
@@ -33,14 +35,19 @@ defmodule SWBenchmark do
     mixed_lines = get_texts("mixed_text")
     long_lines = get_texts("long_text")
 
-    # Empty line for readability
+    # Empty line for readability, sleep to ensure the line is _after_ the logs
+    :timer.sleep(20)
     IO.puts ""
 
     Benchee.run(
-      %{
-        "short" => fn -> compare_all_lines(short_lines) end,
-        "mixed" => fn -> compare_all_lines(mixed_lines) end,
-        "long" => fn -> compare_all_lines(long_lines) end
+      @methods
+      |> Map.new(fn m ->
+        {m, fn lines -> compare_all_lines(lines, m) end}
+      end),
+      inputs: %{
+        "Short" => short_lines,
+        "Mixed" => mixed_lines,
+        "Long" => long_lines
       }
     )
   end