--- a/config/config.exs
+++ b/config/config.exs
@@ -1,3 +1,5 @@
 import Config
 
 config :logger, :console, format: "$message\n"
+
+import_config "#{config_env()}.exs"