--- a/config/test.exs
+++ b/config/test.exs
@@ -1,0 +1,4 @@
+import Config
+
+# Print only warnings and errors during test
+config :logger, level: :warning