--- a/test/main_test.exs
+++ b/test/main_test.exs
@@ -2,21 +2,47 @@ defmodule SmithWaterman.MainTest do
   @moduledoc false
   use ExUnit.Case
 
-  describe "limited testing" do
-    testa = "this is not a test"
-    testb = "this is indeed a test"
-    assert SmithWaterman.score_ratio(testa, testb, 0.7) == 0.6111111111111112
+  describe "lists" do
+    test "short" do
+      subject = "this is not a test"
+      query = "this is indeed a test"
+      assert SmithWaterman.score_ratio(subject, query, 0.7) == 0.6111111111111112
+    end
 
-    testa = "This implementation doesnt perform the traceback, although it could"
+    test "mid" do
+      subject = "This implementation doesnt perform the traceback, although it could"
+      query =
+        "the implementation details perform the perform yeah this text is just here to test things"
 
-    testb =
-      "the implementation details perform the perform yeah this text is just here to test things"
+      assert SmithWaterman.score_ratio(subject, query, 0.7) == 0.13432835820895522
+    end
 
-    assert SmithWaterman.score_ratio(testa, testb, 0.7) == 0.13432835820895522
+    test "identical" do
+      subject = "This implementation doesnt perform the traceback, although it could"
+      query = "This implementation doesnt perform the traceback, although it could"
+      assert SmithWaterman.score_ratio(subject, query, 0.7) == 1.0
+    end
+  end
 
-    # Test identical
-    testa = "This implementation doesnt perform the traceback, although it could"
-    testb = "This implementation doesnt perform the traceback, although it could"
-    assert SmithWaterman.score_ratio(testa, testb, 0.7) == 1.0
+  describe "optimised lists" do
+    test "short" do
+      subject = "this is not a test"
+      query = "this is indeed a test"
+      assert SmithWaterman.score_ratio(subject, query, 0.7, :op_lists) == 0.6111111111111112
+    end
+
+    test "mid" do
+      subject = "This implementation doesnt perform the traceback, although it could"
+      query =
+        "the implementation details perform the perform yeah this text is just here to test things"
+
+      assert SmithWaterman.score_ratio(subject, query, 0.7, :op_lists) == 0.13432835820895522
+    end
+
+    test "identical" do
+      subject = "This implementation doesnt perform the traceback, although it could"
+      query = "This implementation doesnt perform the traceback, although it could"
+      assert SmithWaterman.score_ratio(subject, query, 0.7, :op_lists) == 1.0
+    end
   end
 end