Running `mix benchmark` provides me with the following results.

```
Compiling 1 file (.ex)
Opened file short_text.txt, extracted 12 lines for 144 comparisons
Opened file mixed_text.txt, extracted 6 lines for 36 comparisons
Opened file long_text.txt, extracted 4 lines for 16 comparisons

Operating System: Linux
CPU Information: AMD Ryzen 9 5950X 16-Core Processor
Number of Available Cores: 32
Available memory: 62.77 GB
Elixir 1.14.3
Erlang 25.2.3
JIT enabled: true

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 0 ns
reduction time: 0 ns
parallel: 1
inputs: Long, Mixed, Short
Estimated total run time: 42 s

Benchmarking lists with input Long ...
Benchmarking lists with input Mixed ...
Benchmarking lists with input Short ...
Benchmarking op_lists with input Long ...
Benchmarking op_lists with input Mixed ...
Benchmarking op_lists with input Short ...
Calculating statistics...
Formatting results...

##### With input Long #####
Name               ips        average  deviation         median         99th %
op_lists         10.30       97.09 ms     ±2.97%       96.84 ms      108.58 ms
lists             9.24      108.24 ms     ±0.51%      108.14 ms      109.89 ms

Comparison: 
op_lists         10.30
lists             9.24 - 1.11x slower +11.15 ms

##### With input Mixed #####
Name               ips        average  deviation         median         99th %
op_lists         36.94       27.07 ms     ±0.81%       27.06 ms       28.40 ms
lists            32.61       30.67 ms     ±0.58%       30.65 ms       31.17 ms

Comparison: 
op_lists         36.94
lists            32.61 - 1.13x slower +3.60 ms

##### With input Short #####
Name               ips        average  deviation         median         99th %
op_lists        182.81        5.47 ms     ±5.21%        5.38 ms        6.11 ms
lists           165.31        6.05 ms     ±2.50%        6.00 ms        6.44 ms

Comparison: 
op_lists        182.81
lists           165.31 - 1.11x slower +0.58 ms
```